move to smallint

+++ sql-schema/135.sql
@@ -1,1 +1,1 @@
-ALTER TABLE state_translations MODIFY state_value INTEGER NOT NULL;
+ALTER TABLE state_translations MODIFY state_value SMALLINT(5) NOT NULL;
